project/requirements.txt
Django==1.7.4
Fabric==1.10.1
argparse==1.2.1
django-allauth==0.19.1
django-tastypie==0.12.1
ecdsa==0.13
httplib2==0.9
oauth2==1.5.211
oauthlib==0.7.2
paramiko==1.15.2
pycrypto==2.6.1
python-dateutil==2.4.0
python-mimeparse==0.1.4
python-openid==2.2.5
requests==2.5.1
requests-oauthlib==0.4.2
six==1.9.0
wsgiref==0.1.2
django-haystack==2.3.1
pyelasticsearch==1.1
Pillow==2.7.0
py-moneyed==0.5.0
django-money==0.5.0
